requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
anthropic==0.7.2
pydantic==2.5.1
python-dotenv==1.0.0
requests==2.31.0
pytest==7.4.3
httpx==0.25.2
pytest-asyncio==0.21.1
pytest-cov==4.1.0
coverage==7.3.2
python-hcl2==4.3.2
PyGithub==2.1.1
sqlalchemy==2.0.23
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
semver